# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/calls/active": {
      "get": {
        "operationId": "CallsController_getActiveCalls",
        "summary": "활성 콜 목록",
        "description": "sessionStatus 가 ENDED 가 아닌 테넌트 통화 세션 (최근 100건). agentName·waitSeconds 포함.",
        "parameters": [
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/history": {
      "get": {
        "operationId": "CallsController_listHistory",
        "summary": "통화내역 조회 (CDR)",
        "description": "날짜/상담원/상태 필터, 최근 500건",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ENDED",
                "QUEUED",
                "TALKING",
                "AFTER_CALL_WORK",
                "RINGING_AGENT"
              ],
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "missed",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/recordings/list": {
      "get": {
        "operationId": "CallsController_listRecordings",
        "summary": "녹취 목록 조회",
        "description": "날짜 필터, 최근 200건",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}": {
      "get": {
        "operationId": "CallsController_getCallDetail",
        "summary": "콜 상세",
        "description": "콜 세션 + callLegs + callMemos + callTransfers + callRecordings + queueEvents",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/originate": {
      "post": {
        "operationId": "CallsController_originate",
        "summary": "Click-to-Call 발신 요청",
        "description": "AMI Action:Originate 를 상담원 내선으로 전송. 실제 성공 판정은 후속 DialBegin/DialEnd/BridgeEnter/Newstate 이벤트로 SessionEngine 이 담당하며, 이 응답은 즉시 accepted:true 를 반환한다 (conv 40). Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OriginateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/originate/internal": {
      "post": {
        "operationId": "CallsController_originateInternal",
        "summary": "상담원 간 내선 Click-to-Call 요청",
        "description": "현재 로그인한 상담원 내선을 먼저 울린 뒤, 응답하면 targetExtension 으로 내선 연결을 시도한다. 상대 상담원이 응답하면 통화가 성립한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalOriginateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/transfer": {
      "post": {
        "operationId": "CallsController_transfer",
        "summary": "호 전환 (blind / attended)",
        "description": "legType=agent && !endedAt 인 상담원 leg 의 channelName 을 찾아 AMI Action:Redirect 로 transfer-target context 로 점프 (infra/asterisk/extensions_transfer.conf). BlindTransfer/AttendedTransfer AMI 이벤트 수신 시 TransferDetectorService 가 COMPLETED 로 확정. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/transfer/attended/cancel": {
      "post": {
        "operationId": "CallsController_cancelAttendedTransfer",
        "summary": "상담 전환 취소",
        "description": "열린 attended transfer candidate 를 FAILED/CANCELED 로 닫고, AMI CancelAtxfer 로 취소를 요청한다. 실제 취소 동작은 Asterisk features.conf 의 atxferabort 설정에 의존한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/transfer/attended/complete": {
      "post": {
        "operationId": "CallsController_completeAttendedTransfer",
        "summary": "상담 전환 완료",
        "description": "Asterisk features.conf 의 atxfercomplete 기능 코드를 AMI PlayDTMF 로 제어 채널에 주입한다. 실제 완료 판정은 후속 AttendedTransfer 이벤트 수신 시 TransferDetectorService 가 COMPLETED 로 확정한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/pickup": {
      "post": {
        "operationId": "CallsController_pickup",
        "summary": "대기 콜 당겨받기 요청",
        "description": "큐 대기 중 고객 leg 를 현재 로그인한 상담원 내선으로 Redirect 한다. 실제 연결 성공 판정은 후속 Dial/Bridge 이벤트로 SessionEngine 이 담당하며, 이 응답은 즉시 accepted:true 를 반환한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/mute": {
      "post": {
        "operationId": "CallsController_mute",
        "summary": "통화 음소거/해제 요청",
        "description": "상담원 leg 에 AMI MuteAudio 를 전송한다. 현재 구조에서는 후속 mute 상태 이벤트를 별도로 동기화하지 않으므로, UI는 요청 성공 기준으로 상태를 갱신한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteCallDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/calls/{callId}/hold": {
      "post": {
        "operationId": "CallsController_hold",
        "summary": "통화 hold 요청",
        "description": "표준 AMI hold 액션이 없으므로, 운영자가 검증한 feature code 가 설정된 경우에만 AMI PlayDTMF 로 hold 요청을 전달한다. 실제 상태는 후속 Hold/Unhold 이벤트를 신뢰한다. Clients may send x-correlation-id and idempotency-key headers; the response echoes correlationId and requestedAt.",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-correlation-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAckResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/calls/{callId}/memo": {
      "post": {
        "operationId": "CallsController_memo",
        "summary": "상담 메모 / 후처리 코드 저장",
        "parameters": [
          {
            "name": "callId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "calls"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "KAster CTI API",
    "description": "Asterisk CTI middleware API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "error": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "OriginateDto": {
        "type": "object",
        "properties": {
          "agentExtension": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "callerId": {
            "type": "string",
            "description": "허용된 발신번호 목록에 포함된 번호만 사용 가능"
          }
        },
        "required": [
          "agentExtension",
          "phoneNumber"
        ]
      },
...
      "CommandAckResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/CommandAckDto"
          },
          "error": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
    }
  }
}